+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Bugfixes:
 
   - Use #underscore, not #downcase for help text scope (#140, @atipugin)
   - Radio button and checkbox labels will now include the disabled class as needed. (#156, @ScottSwezey)
+  - Fixed issue with setting offset in form_group without label in horizontal layout  
+    when form uses non-default label_col
 
 Features:
 
@@ -11,6 +13,7 @@ Features:
   - Use humanized attribute name in label errors (#146, @atipugin)
   - Allow to skip label rendering (#145, @atipugin)
   - Added a `required` CSS class for labels with required attributes (#150, @krsyoung)
+  - Added option to customize labels' class.
 
 ## 2.2.0 (2014-09-16)